Locomote fork of global_phone (NB. don't use)
This is a fork of the wonderful sstepenson/global_phone it only exists
because our company has a runtime dependency on this gem as part of a
private gem that we use in a rails app.  And the gemspec for our private
gem can't reference a github branch.  Hopefully we can delete this if
our PRs get merged (ie. sstephenson/global_phone/pull/6 ,
sstephenson/global_phone/pull/7 , sstephenson/global_phone/pull/8 ,
sstephenson/global_phone/pull/9, sstephenson/global_phone/pull/10) and
we can go back to depending on global_phone.